chore(YB-502): 통합테스트 중복 코드를 줄이기 위한 Base 클래스 추가
